# lib/api-zod/src/generated/types/adminActivity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminActivity {
  id: number;
  message: string;
  createdAt: Date;
}