# lib/api-zod/src/generated/types/adminCoin.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminCoinStatus } from "./adminCoinStatus";

export interface AdminCoin {
  id: number;
  symbol: string;
  name: string;
  network: string;
  price: number;
  status: AdminCoinStatus;
  depositEnabled: boolean;
  withdrawalEnabled: boolean;
  createdAt: Date;
}