# lib/api-zod/src/generated/types/adminFeeTier.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminFeeTier {
  id: number;
  name: string;
  makerFee: number;
  takerFee: number;
  withdrawalFee: number;
  minVolume: number;
}