# lib/api-zod/src/generated/types/adminOverview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminOverview {
  listedCoins: number;
  totalCoins: number;
  activePairs: number;
  totalPairs: number;
  users: number;
  pendingKyc: number;
  topMakerFee: number;
}