# lib/api-zod/src/generated/types/adminPair.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminPairStatus } from "./adminPairStatus";

export interface AdminPair {
  id: number;
  base: string;
  quote: string;
  minOrder: number;
  maxLeverage: number;
  status: AdminPairStatus;
  createdAt: Date;
}