# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUserKyc } from "./adminUserKyc";
import type { AdminUserRole } from "./adminUserRole";
import type { AdminUserStatus } from "./adminUserStatus";

export interface AdminUser {
  id: number;
  name: string;
  email: string;
  role: AdminUserRole;
  kyc: AdminUserKyc;
  status: AdminUserStatus;
  balance: number;
  createdAt: Date;
}