# lib/api-zod/src/generated/types/createAdminCoin.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAdminCoin {
  symbol: string;
  name: string;
  network: string;
  price: number;
}