# lib/api-zod/src/generated/types/createAdminUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAdminUserRole } from "./createAdminUserRole";

export interface CreateAdminUser {
  name: string;
  email: string;
  role: CreateAdminUserRole;
}